[#62][hub] refactor: 다익스트라 알고리즘 리펙토링
